@@ -58,79 +58,79 @@
     "mockjs": "^1.1.0",
     "nprogress": "^0.2.0",
     "path": "^0.12.7",
-    "pinia": "^2.2.4",
-    "qs": "^6.13.0",
-    "sortablejs": "^1.15.3",
-    "vue": "^3.5.12",
-    "vue-i18n": "^10.0.4",
-    "vue-router": "^4.4.5",
+    "pinia": "^2.3.0",
+    "qs": "^6.13.1",
+    "sortablejs": "^1.15.6",
+    "vue": "^3.5.13",
+    "vue-i18n": "^10.0.5",
+    "vue-router": "^4.5.0",
     "vue-types": "^5.1.3"
   },
   "devDependencies": {
     "@esbuild-kit/cjs-loader": "^2.4.4",
-    "@jsxiaosi/commitlint-config": "^1.0.5",
-    "@jsxiaosi/eslint-config": "^1.0.5",
-    "@jsxiaosi/eslint-config-prettier": "^1.0.5",
+    "@jsxiaosi/commitlint-config": "^1.0.7",
+    "@jsxiaosi/eslint-config": "^1.0.7",
+    "@jsxiaosi/eslint-config-prettier": "^1.0.7",
     "@types/core-js": "^2.5.8",
     "@types/lodash-es": "^4.17.12",
     "@types/mockjs": "^1.0.10",
     "@types/node": "v22.7.8",
     "@types/nprogress": "^0.2.3",
-    "@types/qs": "^6.9.16",
-    "@vitejs/plugin-vue": "^5.1.4",
-    "@vitejs/plugin-vue-jsx": "^4.0.1",
+    "@types/qs": "^6.9.17",
+    "@vitejs/plugin-vue": "^5.2.1",
+    "@vitejs/plugin-vue-jsx": "^4.1.1",
     "@vue/test-utils": "^2.4.6",
     "autoprefixer": "^10.4.20",
     "conventional-changelog-cli": "^5.0.0",
     "cross-env": "^7.0.3",
-    "czg": "^1.10.1",
+    "czg": "^1.11.0",
     "electron": "^33.0.2",
     "electron-builder": "^25.1.8",
     "electron-reload": "2.0.0-alpha.1",
-    "eslint": "^9.13.0",
+    "eslint": "^9.16.0",
     "eslint-plugin-prettier": "^5.2.1",
-    "eslint-plugin-vue": "^9.29.1",
+    "eslint-plugin-vue": "^9.32.0",
     "http-server": "^14.1.1",
-    "husky": "^9.1.6",
+    "husky": "^9.1.7",
     "jsdom": "^25.0.1",
     "kolorist": "^1.8.0",
-    "lint-staged": "^15.2.10",
-    "postcss": "^8.4.47",
+    "lint-staged": "^15.2.11",
+    "postcss": "^8.4.49",
     "postcss-html": "^1.7.0",
     "postcss-import": "^16.1.0",
     "postcss-scss": "^4.0.9",
-    "prettier": "^3.3.3",
+    "prettier": "^3.4.2",
     "pretty-quick": "^4.0.0",
     "rollup-plugin-visualizer": "^5.12.0",
-    "sass": "^1.79.5",
-    "sass-migrator": "^2.2.0",
-    "stylelint": "^16.10.0",
+    "sass": "^1.82.0",
+    "sass-migrator": "^2.2.1",
+    "stylelint": "^16.11.0",
     "stylelint-config-property-sort-order-smacss": "^10.0.0",
     "stylelint-config-recommended": "^14.0.1",
     "stylelint-config-recommended-scss": "^14.1.0",
     "stylelint-config-recommended-vue": "^1.5.0",
     "stylelint-config-standard": "^36.0.1",
-    "stylelint-config-standard-scss": "^13.1.0",
+    "stylelint-config-standard-scss": "^14.0.0",
     "stylelint-order": "^6.0.4",
-    "stylelint-scss": "^6.8.1",
-    "tailwindcss": "^3.4.14",
+    "stylelint-scss": "^6.10.0",
+    "tailwindcss": "^3.4.16",
     "tree-kill": "^1.2.2",
-    "typescript": "^5.6.3",
+    "typescript": "^5.7.2",
     "unplugin-element-plus": "^0.8.0",
-    "vite": "^5.4.9",
+    "vite": "^6.0.3",
     "vite-plugin-compression": "^0.5.1",
     "vite-plugin-electron": "^0.28.8",
     "vite-plugin-electron-renderer": "^0.14.6",
-    "vite-plugin-fake-server": "^2.1.2",
+    "vite-plugin-fake-server": "^2.1.4",
     "vite-plugin-imagemin": "^0.6.1",
-    "vite-plugin-inspect": "^0.8.7",
-    "vite-plugin-linter": "^3.0.0",
-    "vite-plugin-pwa": "^0.20.5",
+    "vite-plugin-inspect": "^0.10.3",
+    "vite-plugin-linter": "^3.0.1",
+    "vite-plugin-pwa": "^0.21.1",
     "vite-plugin-style-import": "^2.0.0",
     "vite-plugin-svg-icons": "^2.0.1",
-    "vitest": "^2.1.3",
+    "vitest": "^2.1.8",
     "vue-eslint-parser": "^9.4.3",
-    "vue-tsc": "^2.1.6",
+    "vue-tsc": "^2.1.10",
     "xlsx": "^0.18.5"
   },
   "pnpm": {
